requirements.txt
openai>=1.60.0
requests>=2.31.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
jinja2>=3.1.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
